triangle refactoring complete
