+++ geometry/__init__.py
@@ -1,3 +1,4 @@
 from .point import Point
 from .line import Line
 from .circle import Circle
+from .triangle import Triangle
